+++ components/dropzone/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="5.0.9"></a>
+##5.0.9
+🗓
+2023-07-17 • 📝 [Commits](https://github.com/adobe/spectrum-css/compare/@spectrum-css/dropzone@5.0.8...@spectrum-css/dropzone@5.0.9)
+
+**Note:** Version bump only for package @spectrum-css/dropzone
+
+<a name="5.0.8"></a>
+##5.0.8
+🗓
+2023-07-14 • 📝 [Commits](https://github.com/adobe/spectrum-css/compare/@spectrum-css/dropzone@5.0.7...@spectrum-css/dropzone@5.0.8)
+
+**Note:** Version bump only for package @spectrum-css/dropzone
+
 <a name="5.0.7"></a>
 ##5.0.7
 🗓