+++ components/rating/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.0.13"></a>
+##4.0.13
+🗓
+2023-07-17 • 📝 [Commits](https://github.com/adobe/spectrum-css/compare/@spectrum-css/rating@4.0.12...@spectrum-css/rating@4.0.13)
+
+**Note:** Version bump only for package @spectrum-css/rating
+
+<a name="4.0.12"></a>
+##4.0.12
+🗓
+2023-07-14 • 📝 [Commits](https://github.com/adobe/spectrum-css/compare/@spectrum-css/rating@4.0.11...@spectrum-css/rating@4.0.12)
+
+**Note:** Version bump only for package @spectrum-css/rating
+
 <a name="4.0.11"></a>
 ##4.0.11
 🗓